feat: split PathEnvironment into separate module
This is a prep patch for introducing ManagedEnvironment and
RemoteEnvironment structs. Shared functionality will be in a common
module, but each struct will be in its own module.
